+++ circuits/environment/src/helpers/converter.rs
@@ -154,9 +154,9 @@ impl<F: PrimeField> snarkvm_r1cs::ConstraintSynthesizer<F> for R1CS<F> {
         }
 
         // Ensure the given `cs` matches in size with the first system.
-        assert_eq!(self.num_public() + 1, cs.num_public_variables());
-        assert_eq!(self.num_private(), cs.num_private_variables());
-        assert_eq!(self.num_constraints(), cs.num_constraints());
+        assert_eq!(self.num_public() + 1, cs.num_public_variables() as u64);
+        assert_eq!(self.num_private(), cs.num_private_variables() as u64);
+        assert_eq!(self.num_constraints(), cs.num_constraints() as u64);
 
         Ok(())
     }
